+++ CHANGELOG.md
@@ -7,3 +7,11 @@
 - Added ability to add new chats
 - Added ability to send messages to chat
 
+📍 v1.0.1
+- Created message tile design
+- Added user's profile picture to
+  message tile (profile picture only
+  appears on the first message if a
+  user sends multiple messages in a row)
+- Added drawer with chat name and picture
+  as header to chat screen